+++ src/components/Message/__tests__/MessageTeam.test.js
@@ -13,6 +13,7 @@ import {
 import { ChannelContext } from '../../../context';
 import MessageTeam from '../MessageTeam';
 import { Avatar as AvatarMock } from '../../Avatar';
+import { MML as MMLMock } from '../../MML';
 import { MessageInput as MessageInputMock } from '../../MessageInput';
 import { MessageActions as MessageActionsMock } from '../../MessageActions';
 
@@ -27,6 +28,8 @@ jest.mock('../../MessageInput', () => ({
 jest.mock('../../MessageActions', () => ({
   MessageActions: jest.fn(() => <div />),
 }));
+
+jest.mock('../../MML', () => ({ MML: jest.fn(() => <div />) }));
 
 const alice = generateUser({ name: 'alice', image: 'alice-avatar.jpg' });
 const bob = generateUser({ name: 'bob', image: 'bob-avatar.jpg' });
@@ -157,6 +160,31 @@ describe('<MessageTeam />', () => {
     expect(getByTestId('custom-avatar')).toBeInTheDocument();
   });
 
+  it('should render custom edit message input component when one is given', async () => {
+    const message = generateAliceMessage();
+    const updateMessage = jest.fn();
+    const clearEditingState = jest.fn();
+
+    const CustomEditMessageInput = () => <div>Edit Input</div>;
+
+    await renderMessageTeam(message, {
+      clearEditingState,
+      editing: true,
+      updateMessage,
+      EditMessageInput: CustomEditMessageInput,
+    });
+
+    expect(MessageInputMock).toHaveBeenCalledWith(
+      expect.objectContaining({
+        clearEditingState,
+        message,
+        Input: CustomEditMessageInput,
+        updateMessage,
+      }),
+      {},
+    );
+  });
+
   it('should render message input when in edit mode', async () => {
     const message = generateAliceMessage();
     const updateMessage = jest.fn();
@@ -172,6 +200,16 @@ describe('<MessageTeam />', () => {
         updateMessage,
         clearEditingState,
       }),
+      {},
+    );
+  });
+
+  it('should render MML', async () => {
+    const mml = '<mml>text</mml>';
+    const message = generateAliceMessage({ mml });
+    await renderMessageTeam(message);
+    expect(MMLMock).toHaveBeenCalledWith(
+      expect.objectContaining({ source: mml, align: 'left' }),
       {},
     );
   });